Правка бандлов, и основных вызовов класса
