+++ ОВиВК.tab/Спецификации.panel/Дополнительные.stack/Примечания и номера.pulldown/Площади в примечания.pushbutton/bundle.yaml
@@ -4,7 +4,20 @@ title:
 
 tooltip:
   ru: Заполнение площадей воздуховодов в примечания.
+  
+    1. Откройте вид спецификации, на котором требуется заполнить площади в примечания воздуховодов.
+    
+    2. Запустите команду.
+    
+    3. Проверьте результат работы. Параметры типа заполнены не будут!
+
   en_us: Filling duct areas to notes.
+  
+    1. Open the specification view where you want to fill in the areas in the duct notes.
+  
+    2. Run the command.
+  
+    3. Check the result of the work. The type parameters will not be filled in!
 
 author: Redomine
 help_url: "https://kb.a101.ru/pages/viewpage.action?pageId=110562607"
